물품 대여 기능 구현

+++ src/test/java/com/dku/council/domain/like/PostLikeServiceTest.java
@@ -185,7 +185,7 @@ class PostLikeServiceTest {
 
         when(memoryRepository.getAllPostLikes()).thenReturn(likes);
         when(userRepository.getReferenceById(any()))
-                .thenAnswer(inv -> UserMock.create(inv.getArgument(0)));
+                .thenAnswer(inv -> UserMock.create((Long) inv.getArgument(0)));
         when(postRepository.getReferenceById(any()))
                 .thenAnswer(inv -> NewsMock.create(inv.getArgument(0)));
